fix: polish landing download surfaces
